+++ crates/oxc_transformer/src/typescript/namespace.rs
@@ -30,7 +30,7 @@ impl<'a> TypeScript<'a> {
             return;
         }
 
-        // Collect all binding names. Such as function name and class name.
+        // Collect function/class/enum/namespace binding names
         let mut names: FxHashSet<Atom<'a>> = FxHashSet::default();
 
         // Recreate the statements vec for memory efficiency.
@@ -59,12 +59,11 @@ impl<'a> TypeScript<'a> {
                         }
                     }
                     new_stmts.push(Statement::TSModuleDeclaration(decl));
+                    continue;
                 }
-                match_module_declaration!(Statement) => {
-                    if let Statement::ExportNamedDeclaration(export_decl) = &stmt {
-                        if let Some(Declaration::TSModuleDeclaration(decl)) =
-                            &export_decl.declaration
-                        {
+                Statement::ExportNamedDeclaration(ref export_decl) => {
+                    match &export_decl.declaration {
+                        Some(Declaration::TSModuleDeclaration(decl)) => {
                             if !decl.modifiers.is_contains_declare() {
                                 if !self.options.allow_namespaces {
                                     self.ctx.error(namespace_not_supported(decl.span));
@@ -83,53 +82,40 @@ impl<'a> TypeScript<'a> {
                                                 SPAN,
                                                 declaration,
                                             );
-                                        let export_named_decl =
-                                            ModuleDeclaration::ExportNamedDeclaration(
-                                                export_named_decl,
-                                            );
                                         let stmt =
-                                            self.ctx.ast.module_declaration(export_named_decl);
+                                            Statement::ExportNamedDeclaration(export_named_decl);
                                         new_stmts.push(stmt);
                                     }
                                     new_stmts.push(transformed_stmt);
                                     continue;
                                 }
                             }
-                        }
-                    }
 
-                    stmt.to_module_declaration().bound_names(&mut |id| {
-                        names.insert(id.name.clone());
-                    });
-                    new_stmts.push(stmt);
-                }
-                // Collect bindings from class, function, variable and enum declarations
-                Statement::FunctionDeclaration(ref decl) => {
-                    if let Some(ident) = &decl.id {
-                        names.insert(ident.name.clone());
+                            if let TSModuleDeclarationName::Identifier(id) = &decl.id {
+                                names.insert(id.name.clone());
+                            }
+                        }
+                        Some(decl) => match decl {
+                            Declaration::FunctionDeclaration(_)
+                            | Declaration::ClassDeclaration(_)
+                            | Declaration::TSEnumDeclaration(_) => {
+                                names.insert(decl.id().as_ref().unwrap().name.clone());
+                            }
+                            _ => {}
+                        },
+                        _ => {}
                     }
-                    new_stmts.push(stmt);
                 }
-                Statement::ClassDeclaration(ref decl) => {
-                    if let Some(ident) = &decl.id {
-                        names.insert(ident.name.clone());
-                    }
-                    new_stmts.push(stmt);
+                // Collect bindings from class, function and enum declarations
+                Statement::FunctionDeclaration(_)
+                | Statement::ClassDeclaration(_)
+                | Statement::TSEnumDeclaration(_) => {
+                    names.insert(stmt.to_declaration().id().as_ref().unwrap().name.clone());
                 }
-                Statement::TSEnumDeclaration(ref decl) => {
-                    names.insert(decl.id.name.clone());
-                    new_stmts.push(stmt);
-                }
-                Statement::VariableDeclaration(ref decl) => {
-                    decl.bound_names(&mut |id| {
-                        names.insert(id.name.clone());
-                    });
-                    new_stmts.push(stmt);
-                }
-                _ => {
-                    new_stmts.push(stmt);
-                }
+                _ => {}
             }
+
+            new_stmts.push(stmt);
         }
 
         program.body = new_stmts;
@@ -141,39 +127,37 @@ impl<'a> TypeScript<'a> {
         parent_export: Option<Expression<'a>>,
         ctx: &mut TraverseCtx,
     ) -> Option<Statement<'a>> {
+        // Skip empty declaration e.g. `namespace x;`
+        let body = decl.body?;
+
         let mut names: FxHashSet<Atom<'a>> = FxHashSet::default();
 
-        let real_name = decl.id.name();
+        let TSModuleDeclarationName::Identifier(IdentifierName { name: real_name, .. }) = decl.id
+        else {
+            return None;
+        };
 
-        // TODO: This binding is created in wrong scope.
-        // Needs to be created in scope of function which `transform_namespace` creates below.
-        let name = self.ctx.ast.new_atom(
-            &ctx.generate_uid_in_current_scope(real_name, SymbolFlags::FunctionScopedVariable),
-        );
+        // Reuse `TSModuleDeclaration`'s scope in transformed function
+        let scope_id = decl.scope_id.get().unwrap();
+        let name = self.ctx.ast.new_atom(&ctx.generate_uid(
+            &real_name,
+            scope_id,
+            SymbolFlags::FunctionScopedVariable,
+        ));
 
-        // Reuse TSModuleBlock's scope id in transformed function.
-        let mut scope_id = None;
-
-        let namespace_top_level = if let Some(body) = decl.body {
-            match body {
-                TSModuleDeclarationBody::TSModuleBlock(block) => {
-                    scope_id = block.scope_id.get();
-                    block.unbox().body
-                }
-                // We handle `namespace X.Y {}` as if it was
-                //   namespace X {
-                //     export namespace Y {}
-                //   }
-                TSModuleDeclarationBody::TSModuleDeclaration(declaration) => {
-                    let declaration = Declaration::TSModuleDeclaration(declaration);
-                    let export_named_decl =
-                        self.ctx.ast.plain_export_named_declaration_declaration(SPAN, declaration);
-                    let stmt = Statement::ExportNamedDeclaration(export_named_decl);
-                    self.ctx.ast.new_vec_single(stmt)
-                }
+        let namespace_top_level = match body {
+            TSModuleDeclarationBody::TSModuleBlock(block) => block.unbox().body,
+            // We handle `namespace X.Y {}` as if it was
+            //   namespace X {
+            //     export namespace Y {}
+            //   }
+            TSModuleDeclarationBody::TSModuleDeclaration(declaration) => {
+                let declaration = Declaration::TSModuleDeclaration(declaration);
+                let export_named_decl =
+                    self.ctx.ast.plain_export_named_declaration_declaration(SPAN, declaration);
+                let stmt = Statement::ExportNamedDeclaration(export_named_decl);
+                self.ctx.ast.new_vec_single(stmt)
             }
-        } else {
-            self.ctx.ast.new_vec()
         };
 
         let mut new_stmts = self.ctx.ast.new_vec();
@@ -183,6 +167,7 @@ impl<'a> TypeScript<'a> {
                 Statement::TSModuleDeclaration(decl) => {
                     if decl.id.is_string_literal() {
                         self.ctx.error(ambient_module_nested(decl.span));
+                        continue;
                     }
 
                     let module_name = decl.id.name().clone();
@@ -194,28 +179,7 @@ impl<'a> TypeScript<'a> {
                         }
                         new_stmts.push(transformed);
                     }
-                }
-                Statement::ClassDeclaration(ref decl) => {
-                    decl.bound_names(&mut |id| {
-                        names.insert(id.name.clone());
-                    });
-                    new_stmts.push(stmt);
-                }
-                Statement::FunctionDeclaration(ref decl) => {
-                    decl.bound_names(&mut |id| {
-                        names.insert(id.name.clone());
-                    });
-                    new_stmts.push(stmt);
-                }
-                Statement::TSEnumDeclaration(ref enum_decl) => {
-                    names.insert(enum_decl.id.name.clone());
-                    new_stmts.push(stmt);
-                }
-                Statement::VariableDeclaration(ref decl) => {
-                    decl.bound_names(&mut |id| {
-                        names.insert(id.name.clone());
-                    });
-                    new_stmts.push(stmt);
+                    continue;
                 }
                 Statement::ExportNamedDeclaration(export_decl) => {
                     // NB: `ExportNamedDeclaration` with no declaration (e.g. `export {x}`) is not
@@ -243,6 +207,7 @@ impl<'a> TypeScript<'a> {
                             Declaration::TSModuleDeclaration(module_decl) => {
                                 if module_decl.id.is_string_literal() {
                                     self.ctx.error(ambient_module_nested(module_decl.span));
+                                    continue;
                                 }
 
                                 let module_name = module_decl.id.name().clone();
@@ -264,16 +229,20 @@ impl<'a> TypeScript<'a> {
                             _ => {}
                         }
                     }
+                    continue;
                 }
-                stmt => {
-                    if let Some(decl) = stmt.as_declaration() {
-                        if decl.is_typescript_syntax() {
-                            continue;
-                        }
-                    }
-                    new_stmts.push(stmt);
+                // Collect bindings from class, function and enum declarations
+                Statement::ClassDeclaration(_)
+                | Statement::FunctionDeclaration(_)
+                | Statement::TSEnumDeclaration(_) => {
+                    names.insert(stmt.to_declaration().id().as_ref().unwrap().name.clone());
                 }
+                Statement::TSTypeAliasDeclaration(_)
+                | Statement::TSInterfaceDeclaration(_)
+                | Statement::TSImportEqualsDeclaration(_) => continue,
+                _ => {}
             }
+            new_stmts.push(stmt);
         }
 
         if new_stmts.is_empty() {
@@ -285,14 +254,14 @@ impl<'a> TypeScript<'a> {
             return None;
         }
 
-        Some(self.transform_namespace(&name, real_name, new_stmts, parent_export, scope_id, ctx))
+        Some(self.transform_namespace(name, real_name, new_stmts, parent_export, scope_id, ctx))
     }
 
     // `namespace Foo { }` -> `let Foo; (function (_Foo) { })(Foo || (Foo = {}));`
     //                         ^^^^^^^
     fn create_variable_declaration(&self, name: &Atom<'a>) -> Declaration<'a> {
         let kind = VariableDeclarationKind::Let;
-        let declarator = {
+        let declarations = {
             let ident = BindingIdentifier::new(SPAN, name.clone());
             let pattern_kind = self.ctx.ast.binding_pattern_identifier(ident);
             let binding = self.ctx.ast.binding_pattern(pattern_kind, None, false);
@@ -302,20 +271,21 @@ impl<'a> TypeScript<'a> {
         Declaration::VariableDeclaration(self.ctx.ast.variable_declaration(
             SPAN,
             kind,
-            declarator,
+            declarations,
             Modifiers::empty(),
         ))
     }
 
     // `namespace Foo { }` -> `let Foo; (function (_Foo) { })(Foo || (Foo = {}));`
     //                                  ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+    #[allow(clippy::needless_pass_by_value)]
     fn transform_namespace(
         &self,
-        arg_name: &Atom<'a>,
-        real_name: &Atom<'a>,
+        arg_name: Atom<'a>,
+        real_name: Atom<'a>,
         stmts: Vec<'a, Statement<'a>>,
         parent_export: Option<Expression<'a>>,
-        scope_id: Option<ScopeId>,
+        scope_id: ScopeId,
         ctx: &mut TraverseCtx,
     ) -> Statement<'a> {
         // `(function (_N) { var x; })(N || (N = {}))`;
@@ -323,10 +293,8 @@ impl<'a> TypeScript<'a> {
         let callee = {
             let body = self.ctx.ast.function_body(SPAN, self.ctx.ast.new_vec(), stmts);
             let params = {
-                let ident = self.ctx.ast.binding_pattern_identifier(BindingIdentifier::new(
-                    SPAN,
-                    self.ctx.ast.new_atom(arg_name),
-                ));
+                let ident =
+                    self.ctx.ast.binding_pattern_identifier(BindingIdentifier::new(SPAN, arg_name));
                 let pattern = self.ctx.ast.binding_pattern(ident, None, false);
                 let items =
                     self.ctx.ast.new_vec_single(self.ctx.ast.plain_formal_parameter(SPAN, pattern));
@@ -344,11 +312,9 @@ impl<'a> TypeScript<'a> {
                 params,
                 Some(body),
             );
-            if let Some(scope_id) = scope_id {
-                function.scope_id.set(Some(scope_id));
-                *ctx.scopes_mut().get_flags_mut(scope_id) =
-                    ScopeFlags::Function | ScopeFlags::StrictMode;
-            }
+            function.scope_id.set(Some(scope_id));
+            *ctx.scopes_mut().get_flags_mut(scope_id) =
+                ScopeFlags::Function | ScopeFlags::StrictMode;
             let function_expr = self.ctx.ast.function_expression(function);
             self.ctx.ast.parenthesized_expression(SPAN, function_expr)
         };
@@ -371,7 +337,7 @@ impl<'a> TypeScript<'a> {
                         self.ctx.ast.static_member(
                             SPAN,
                             parent_export,
-                            self.ctx.ast.identifier_name(SPAN, real_name),
+                            IdentifierName::new(SPAN, real_name.clone()),
                             false,
                         ),
                     )
@@ -397,7 +363,7 @@ impl<'a> TypeScript<'a> {
                     IdentifierReference::new(SPAN, real_name.clone()),
                 );
                 let assign_right = {
-                    let property = self.ctx.ast.identifier_name(SPAN, real_name);
+                    let property = IdentifierName::new(SPAN, real_name.clone());
                     let logical_left =
                         self.ctx.ast.static_member_expression(SPAN, parent_export, property, false);
                     let op = LogicalOperator::Or;