+++ tasks/transform_conformance/src/constants.rs
@@ -91,5 +91,7 @@ pub(crate) const SKIP_TESTS: &[&str] = &[
     "typescript/test/fixtures/namespace/nested-shorthand-export/input.ts",
     "react-jsx-development/test/fixtures/cross-platform/self-inside-arrow/input.mjs",
     // Babel outputs is not correct
+    "typescript/test/fixtures/namespace/clobber-import/input.ts",
+    "typescript/test/fixtures/namespace/namespace-nested-module/input.ts",
     "typescript/test/fixtures/namespace/nested-destructuring/input.ts",
 ];